+++ src/utils-df.h
@@ -1,0 +1,3 @@
+void set_df_attributes(SEXP df_);
+SEXP create_named_list(int n, ...);
+SEXP get_df_col(SEXP df_, const char *str);