+++ NEWS.md
@@ -1,7 +1,9 @@
 
-# zap 0.1.1.9000
+# zap 0.1.1.9001
 
-* [refactor] 2025-07-12 tidy VECSXP to separate functions for 'raw'
+* [9001] [enhance] 2025-07-12 return a data.structure detailing the objects
+  which have been serialized.
+* [9000] [refactor] 2025-07-12 tidy VECSXP to separate functions for 'raw'
   and 'reference' handling.
 
 # zap 0.1.1  2025-07-11